+++ Crypto/Controller/HomeViewController.swift
@@ -11,7 +11,7 @@ class HomeViewController: UIViewController, UITableViewDelegate {
 
     private let APIservice = APIService()
     private var homeView = HomeView()
-    private var coinDS = CoinDataSource()
+    private var coinDataSource = CoinDataSource()
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -21,7 +21,7 @@ class HomeViewController: UIViewController, UITableViewDelegate {
         
         homeView.segmentedControllDelegate = self
         homeView.myTableView.delegate = self
-        homeView.myTableView.dataSource = coinDS
+        homeView.myTableView.dataSource = coinDataSource
     }
     
     override func loadView() {
@@ -33,9 +33,9 @@ class HomeViewController: UIViewController, UITableViewDelegate {
 
 extension HomeViewController: SortingSegmentedControllDelegate {
     func didTypeCoinName(text: String) {
-        var coinArray = coinDS.getCoinArray()
-        var coinArraySearchList = coinDS.getCoinArraySearchList()
-        var coinArrayIndex = coinDS.getCoinArrayIndex()
+        var coinArray = coinDataSource.getCoinArray()
+        var coinArraySearchList = coinDataSource.getCoinArraySearchList()
+        var coinArrayIndex = coinDataSource.getCoinArrayIndex()
         
         if coinArraySearchList.count == 0 {
             coinArraySearchList.append(coinArray)
@@ -52,15 +52,15 @@ extension HomeViewController: SortingSegmentedControllDelegate {
             coinArraySearchList.append(coinArray)
         }
         
-        coinDS.updateCoinArray(array: coinArraySearchList[coinArrayIndex])
-        coinDS.updateCoinArraySearchList(coinArrList: coinArraySearchList)
-        coinDS.updateCoinArrayIndex(indexVal: coinArrayIndex)
+        coinDataSource.updateCoinArray(array: coinArraySearchList[coinArrayIndex])
+        coinDataSource.updateCoinArraySearchList(coinArrList: coinArraySearchList)
+        coinDataSource.updateCoinArrayIndex(indexVal: coinArrayIndex)
         
         homeView.myTableView.reloadData()
     }
     
     func didSelectSegement(segmentIndex: Int) {
-        var coinArray = coinDS.getCoinArray()
+        var coinArray = coinDataSource.getCoinArray()
         switch segmentIndex {
         case 0:
             coinArray = coinArray.sorted(by: {$0.quote.USD.price > $1.quote.USD.price})
@@ -78,55 +78,16 @@ extension HomeViewController: SortingSegmentedControllDelegate {
             print("NONE")
         }
         
-        coinDS.updateCoinArray(array: coinArray)
+        coinDataSource.updateCoinArray(array: coinArray)
         homeView.myTableView.reloadData()
     }
 }
 
 extension HomeViewController: APIProtocol {
-    func dataRetrieved(_ retrievedData: [Coin]) {
-        coinDS.updateCoinArray(array: retrievedData)
+    func coinArrayDidRetrieve(_ retrievedCoinArray: [Coin]) {
+        coinDataSource.updateCoinArray(array: retrievedCoinArray)
         homeView.myTableView.reloadData()
     }
 }
 
-class CoinDataSource: NSObject, UITableViewDataSource {
-    private var coinArray = [Coin]()
-    private var coinArrayIndex: Int = 0
-    private var coinArraySearchList = [[Coin]]()
-    
-    func getCoinArrayIndex() -> Int {
-        return self.coinArrayIndex
-    }
-    
-    func getCoinArraySearchList() -> [[Coin]] {
-        return self.coinArraySearchList
-    }
-    
-    func getCoinArray() -> [Coin] {
-        return self.coinArray
-    }
-    
-    func updateCoinArrayIndex(indexVal: Int) {
-        self.coinArrayIndex = indexVal
-    }
-    
-    func updateCoinArraySearchList(coinArrList: [[Coin]]) {
-        self.coinArraySearchList = coinArrList
-    }
-    
-    func updateCoinArray(array: [Coin]) {
-        self.coinArray = array
-    }
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return coinArray.count
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: CoinTableViewCell.identifier, for: indexPath as IndexPath) as! CoinTableViewCell
-        cell.configureCell(withCell: coinArray[indexPath.row])
-        return cell
-    }
-    
-}
+
